[FEAT] UseCase 구현
