--- domain/src/main/java/com/nexters/knownknowns/domain/BottomSheetUseCase.kt
@@ -1,4 +1,0 @@
-package com.nexters.knownknowns.domain
-
-class BottomSheetUseCase {
-}
